requirement.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
pillow==11.0.0
psycopg==3.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
vobject==0.9.8
